move primer* and string to unused inside qio package
