+++ qio/src/org/qcmg/unused/PrimerDesignRecord.java
@@ -2,7 +2,7 @@
  * © Copyright The University of Queensland 2010-2014.  This code is released under the terms outlined in the included LICENSE file.
  */
 
-package org.qcmg.primerdesignsummary;
+package org.qcmg.unused;
 
 public class PrimerDesignRecord {
