+++ qio/src/org/qcmg/unused/PrimerDesignRecordSerializer.java
@@ -1,7 +1,7 @@
 /**
  * © Copyright The University of Queensland 2010-2014.  This code is released under the terms outlined in the included LICENSE file.
  */
-package org.qcmg.primerdesignsummary;
+package org.qcmg.unused;
 
 import java.io.BufferedReader;
 import java.io.IOException;
