+++ qio/src/org/qcmg/unused/PrimerInputFileWriter.java
@@ -1,7 +1,7 @@
 /**
  * © Copyright The University of Queensland 2010-2014.  This code is released under the terms outlined in the included LICENSE file.
  */
-package org.qcmg.primeroutput;
+package org.qcmg.unused;
 
 import java.io.Closeable;
 import java.io.File;
@@ -9,19 +9,17 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.OutputStream;
 
-import org.qcmg.primeroutput.PrimerOutputRecord;
-
-public final class PrimerOutputFileWriter implements Closeable {
+public final class PrimerInputFileWriter implements Closeable {
 	private static final String EQUALS = "=";
-	private static final PrimerOutputRecordSerializer serializer = new PrimerOutputRecordSerializer();
+	private static final PrimerInputRecordSerializer serializer = new PrimerInputRecordSerializer();
 	private final OutputStream outputStream;
 
-	public PrimerOutputFileWriter(final File file) throws Exception {
+	public PrimerInputFileWriter(final File file) throws Exception {
 		OutputStream stream = new FileOutputStream(file);
 		outputStream = stream;
 	}
 
-	public void add(final PrimerOutputRecord record) throws Exception {
+	public void add(final PrimerInputRecord record) throws Exception {
 		String encoded = serializer.serialise(record);
 		outputStream.write(encoded.getBytes());
 		outputStream.flush();
