+++ qio/src/org/qcmg/unused/PrimerInputRecord.java
@@ -1,7 +1,7 @@
 /**
  * © Copyright The University of Queensland 2010-2014.  This code is released under the terms outlined in the included LICENSE file.
  */
-package org.qcmg.primerinput;
+package org.qcmg.unused;
 
 
 public class PrimerInputRecord {
