+++ qio/src/org/qcmg/unused/PrimerOutputFileWriter.java
@@ -1,7 +1,7 @@
 /**
  * © Copyright The University of Queensland 2010-2014.  This code is released under the terms outlined in the included LICENSE file.
  */
-package org.qcmg.primerinput;
+package org.qcmg.unused;
 
 import java.io.Closeable;
 import java.io.File;
@@ -9,17 +9,19 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.OutputStream;
 
-public final class PrimerInputFileWriter implements Closeable {
+import org.qcmg.unused.PrimerOutputRecord;
+
+public final class PrimerOutputFileWriter implements Closeable {
 	private static final String EQUALS = "=";
-	private static final PrimerInputRecordSerializer serializer = new PrimerInputRecordSerializer();
+	private static final PrimerOutputRecordSerializer serializer = new PrimerOutputRecordSerializer();
 	private final OutputStream outputStream;
 
-	public PrimerInputFileWriter(final File file) throws Exception {
+	public PrimerOutputFileWriter(final File file) throws Exception {
 		OutputStream stream = new FileOutputStream(file);
 		outputStream = stream;
 	}
 
-	public void add(final PrimerInputRecord record) throws Exception {
+	public void add(final PrimerOutputRecord record) throws Exception {
 		String encoded = serializer.serialise(record);
 		outputStream.write(encoded.getBytes());
 		outputStream.flush();
